msm: kgsl: Use DMA APIs for memory pool cache maintenance
After allocating and zeroing pages from system memory or the pool we need
to ensure that the cache is synchronized so that it doesn't cause problems
down the road.  Use dma_sync_sg_for_device to make sure the allocated pages
are clean. This isn't the best way to handle this but we haven't yet come
up with a better way and this does the job.

[Reina: Adapt for 4.19]

Change-Id: I155015fdfc58d0075f17afb8b777f7887644c0fa

Signed-off-by: RyuujiX <saputradenny712@gmail.com>