BACKPORT: soc: qcom: smp2p: Add memory barrier for irq_pending
There is a very tight race where the irq_retrigger function is run
on one cpu and the actual retrigger softirq is running on a second
cpu. When this happens, there may be a chance that the second cpu
will not see the updated irq_pending value from first cpu.

Add a memory barrier to ensure that irq_pending is read correctly.

[ Tashar02: Backport to k4.19 ]

Change-Id: I4bee6427489cf59996a34363dcbb5c28900aa9af

Signed-off-by: RyuujiX <saputradenny712@gmail.com>