crypto: arm64 - Use modern annotations for assembly functions
In an effort to clarify and simplify the annotation of assembly functions
in the kernel new macros have been introduced. These replace ENTRY and
ENDPROC and also add a new annotation for static functions which previously
had no ENTRY equivalent. Update the annotations in the crypto code to the
new macros.

There are a small number of files imported from OpenSSL where the assembly
is generated using perl programs, these are not currently annotated at all
and have not been modified.

Acked-by: Ard Biesheuvel <ardb@kernel.org>
Change-Id: I9a129ea0d068551f39499daa61a03d5a1407bc4e
Signed-off-by: RyuujiX <saputradenny712@gmail.com>