kprofiles: Introduce 2 new API for changing mode on custom events
This commit introduces two new APIs for other drivers
to request a profile mode change. So this new APIs gives us the flexibility to
change profile on any custom event.

kprofiles_set_mode_rollback() : This function takes 2 parameters as input, 1st is the profile mode that it will switch to
and other for the amount of time (in milliseconds) to stay at that mode before switching back to the previously active mode.

kprofiles_set_mode() : This function is rather simple compared to the previous function, it only takes 1 input parameter,
which is the mode to switch the profile to, the mode will not change until this function is manually used again to change the mode.

Signed-off-by: Dakkshesh <dakkshesh5@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>