thaw_fingerprintd: Use strncmp() instead of memcmp()
memcmp() requires specific filename, and the fingerprintd hal can be upgraded to a newer version which will require us to modify the code again, hence let's switch to strncmp().

This makes the code more efficient in a way as we're using sizeof() which is determined at compile-time, so the compiler will calculate the size of the string and make it a constant.

Suggested-by: Justin Crawford <Justin@stacksmash.net>
Change-Id: I449a93ff6619df22209ddc0a52f4e31067d70675
Signed-off-by: RyuujiX <saputradenny712@gmail.com>